[Feat] Medetation Controller 이미지 요ã청 수정중 230921 1745

+++ mt/src/main/java/com/project/mt/meditation/dto/request/MeditationRequestDto.java
@@ -9,8 +9,7 @@ import lombok.Data;
 @Data
 public class MeditationRequestDto {
 
-	private int memberIdx;
-
+	private Long memberIdx;
 	private List<MultipartFile> images;
 
 }
